+++ app/Models/Ingredient.php
@@ -127,6 +127,9 @@ class Ingredient extends Model implements SiteSearchable
             'image_url' => $this->getMainImageUrl(),
             'description' => $this->description,
             'category' => $this->category->name,
+            'strength_abv' => $this->strength,
+            'color' => $this->color ?? 'No color',
+            'origin' => $this->origin ?? 'No origin',
         ];
     }
 }