+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ## Changes
 - Increased cocktail thumbnail size to 400x400
+- Updated ingredients index
+    - Added `strength_abv`, `color`, `origin`
 
 # v1.1.0
 This update includes changes to authorization. Users can now be admins and have elevated privilages. Since this is a new change you will not have any admins in your instance if you are not starting from a fresh install. To make yourself admin you need to run the following command with the email you are using to login: